% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time_lag_matrix}
\alias{time_lag_matrix}
\title{time_lag_matrix}
\usage{
time_lag_matrix(x, maxlag)
}
\arguments{
\item{x}{A matrix of predictors, where the rows are the samples and the columns are the predictors}

\item{maxlag}{maximum time lag variable.}
}
\description{
Build the time lag matrix for the data matrix x.
}

